refactor: move encoding to a separate file
